refactor(pi-exa): split monolithic extension into modular structure
Break down the 718-line extensions/index.ts into focused modules to
prepare for PRD-002 (full Exa API alignment) implementation:

- config.ts: config loading, auth resolution, tool enablement
- formatters.ts: response types and result formatters
- schemas.ts: TypeBox parameter schemas
- web-search.ts: performWebSearch
- web-fetch.ts: performWebFetch
- web-search-advanced.ts: performAdvancedSearch
- index.ts: entry point with tool registration and re-exports

Also fix pre-existing test failures in extension.test.ts by isolating
tests from the real filesystem (HOME/cwd sandboxed to temp dirs) so
loadConfig() is not affected by user settings.
